MODIFICATION DE LA STRUCTURE DU PROJET:
- modularisation  du projet (chaque phase a son propre fichier, conçu avec la GUI en tête)

- nouveauté : outil générateur d'utilisateurs

- plusieurs améliorations au niveau du code, robustesse, qualité, documentation (WIP)
