+++ .idea/encodings.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/users/07706024.txt" charset="windows-1252" />
+    <file url="file://$PROJECT_DIR$/users/37050047.txt" charset="windows-1252" />
+    <file url="PROJECT" charset="windows-1252" />
+  </component>
+</project>
