+++ gen_users/dossier_utilisateurs/users/07272359.txt
@@ -1,0 +1,12 @@
+CPT*Compte A
+CPT*Compte B
+CPT*Compte C
+OPE*30/01/2022*essence*Compte C*-78.4*CB*True*divers
+OPE*20/01/2022*restaurant*Compte C*-95.85*VIR*True*alimentation
+OPE*05/01/2022*restaurant*Compte A*-63.18*VIR*True*divers
+OPE*19/01/2022*vente*Compte A*258.69*CHE*False*transport
+OPE*18/01/2022*pharmacie*Compte C*-84.53*CB*True*santé
+OPE*11/01/2022*salaire*Compte C*62.86*CB*True*sorties
+OPE*16/01/2022*vente*Compte C*163.3*CB*True*divertissement
+BUD*Autres*0*Autres
+BUD*sorties*500*Compte A
