+++ src/cryptage_decryptage.py
@@ -1,0 +1,55 @@
+# -*- coding: utf-8 -*-
+#   cryptage_decryptage.py
+#   |--------------------------------------------|   #
+#   |--------Gestion de Banque (avec IHM)--------|   #
+#   |--Groupe 2 (MOUSSA, ASSEMAT, JIN, ZAMOURI)--|   #
+#   |--------Fonctions crypter, décrypter--------|   #
+#   |--------------------------------------------|   #
+# --Imports-- #
+
+# --Constantes-- #
+CLE_CRYPTAGE = 23
+
+
+# --Fonctions-- #
+
+def cryptage(chaine: str, cle: int) -> str:
+    """
+    Fonction, avec 2 options en paramètres, qui renvoie une chaîne de caractères cryptée
+    avec la méthode César, selon la clé fournie.
+
+    Args:
+        chaine (str): texte à crypter
+        cle (int): clé qui sera utilisée pour le cryptage
+
+    Returns:
+        str: la chaine cryptée
+    """
+    crypte = ''
+    for char in chaine:
+        if char == '\n' or char == '*':
+            crypte += char
+        else:
+            crypte += chr(ord(char) + cle)
+    return crypte
+
+
+def decryptage(chaine: str, cle: int) -> str:
+    """
+    Fonction, avec 2 options en paramètres, qui renvoie une chaîne de caractères décryptée
+    avec la méthode César, selon la clé fournie.
+
+    Args:
+        chaine (str): texte à décrypter
+        cle (int): clé qui sera utilisée pour le décryptage
+
+    Returns:
+        str: la chaine décryptée
+    """
+    decrypte = ''
+    for char in chaine:
+        if char == '\n' or char == '*':
+            decrypte += char
+        else:
+            decrypte += chr(ord(char) - cle)
+    return decrypte
