+++ src/ident.txt
@@ -1,5 +1,15 @@
-IJKLMNOP*HGKLPO*gx*IK
-HPIOJNKM*OMKIKM*dx|*GN
-ONMLKJIH*PONMLK*a|x*HL
-JKLMNOPG*HIJKLM*Xz|*GP
-LMNOPGHI*MLKJIH*cz*HO
+NHGJGOHN*KLKHPM*c*GK
+NIKJIHGL*IIHKJP*c*GL
+LJNJPPLM*KLPOGP*`x|*GK
+KHMHPLLO*OPLOPL*Xx*GO
+GNINIJLP*GPOHLK*\x*IH
+LLKKPGNI*LMOINK*gx*HO
+NPNKOJHL*NLLPGM*Z|*II
+HKHNMMKI*MIJHIM*Zx|*IG
+PIHLPOKG*KNHIPI*dx|*GP
+HINIGHJO*PKOGNP*c|x*II
+MNNHNMPP*PKGHGH*p|}*HI
+HJNJPKHJ*IOONIL*Zx|*HJ
+PMLNOMLK*LMHPJH*Z|*GM
+JNLPMLLO*MGLLHJ*Z|*IG
+OKHMIIJN*PJGMLK*Xx*HI
