+++ users/07272359.txt
@@ -1,0 +1,12 @@
+Xei*Xz5V
+Xei*Xz5W
+Xei*Xz5X
+deZ*HEDEFDGEGG*zzxz*Xz5X*BLMCI*XW*iz*y~z
+deZ*GEDEFDGEGG*zvv*Xz5X*BNJCMJ*k^g*iz*v~zv~
+deZ*EJDEFDGEGG*zvv*Xz5V*BKHCFM*k^g*iz*y~z
+deZ*FNDEFDGEGG*zz*Xz5V*GJMCKN*X]Z*[vz*v
+deZ*FMDEFDGEGG*}vvx~z*Xz5X*BMICJH*XW*iz*vþ
+deZ*FFDEFDGEGG*vv~z*Xz5X*KGCMK*XW*iz*~z
+deZ*FKDEFDGEGG*zz*Xz5X*FKHCH*XW*iz*y~z~zz
+WjY*Vz*E*Vz
+WjY*~z*JEE*Xz5V
