+++ users/12720138.txt
@@ -1,0 +1,17 @@
+Yfj*Y{6W
+Yfj*Y{6X
+ef[*GIEFGEHFHH*{*Y{6X*GKJDH*l_h*j{*w{w
+ef[*GJEFGEHFHH*y{*Y{6W*CKKDFM*Y^[*\w{*w
+ef[*FNEFGEHFHH*{{*Y{6W*GOIDGI*l_h*j{*w{w
+ef[*HOEFGEHFHH*~wwy{*Y{6W*CNDIM*Y^[*\w{*z{{{
+ef[*HHEFGEHFHH*{{*Y{6X*NLDLN*l_h*\w{*w
+ef[*FKEFGEHFHH*{{*Y{6X*GIGDMJ*Y^[*j{*z{{{
+ef[*FHEFGEHFHH*{{y{*Y{6X*CLLDLG*l_h*j{*z{{{
+ef[*GHEFGEHFHH*{{*Y{6X*GMLDLL*l_h*\w{*wÿ
+ef[*HHEFGEHFHH*{*Y{6X*HNFDOI*l_h*\w{*w{w
+ef[*GHEFGEHFHH*{x{{*Y{6X*HHMDKK*Y^[*\w{*w{w
+ef[*HLEFGEHFHH*{*Y{6X*GKJDKO*Y^[*j{*{
+ef[*FNEFGEHFHH*ww{*Y{6W*NKDF*l_h*\w{*z{
+ef[*IGEFGEHFHH*{{*Y{6X*HNGDK*l_h*\w{*w
+XkZ*W{*F*W{
+XkZ*{*KFF*Y{6W
