--- users/19283746.txt
@@ -1,8 +1,0 @@
-JW[*Jvtw{l'Joïx|lz
-JW[*WLS
-VWL*796786979;*Ylz{h|yhu{*Jvtw{l'Joïx|lz*:75<7*JI*[y|l*zvy{plz
-VWL*7?6786979;*Hivuultlu{'Ul{msp*Jvtw{l'Joïx|lz*8<5@@*JI*Mhszl*kp}ly{pzzltlu{
-VWL*8<6786979;*Ðwhynul'Tluz|lssl*Jvtw{l'Joïx|lz*977577*]PY*[y|l*ðwhynul
-I\K*zvy{plz*9<7577*Jvtw{l'Joïx|lz
-I\K*kp}ly{pzzltlu{*877577*Jvtw{l'Joïx|lz
-I\K*ðwhynul*<77577*WLS
