--- users/23456789.txt
@@ -1,27 +1,0 @@
-[hl*[}8[y
-[hl*d}8Y
-[hl*]|}
-[hl*[}8}~y
-[hl*ey|
-gh]*ILGHJGJHJK*g}}*[}8}~y*OHFH*}}*l}*āy}
-gh]*HIGHIGJHJL*k}y{ā*[}8[y*LMFK*[Z*l}*y}y
-gh]*HMGHIGJHJL*[āy*[}8[y*IJFH*[Z*l}*}
-gh]*IHGHIGJHJL*n}}8áy}*[}8[y*IHHFH*naj*l}*āy}
-gh]*JHGHQGJHJL*Y}y}*[}8[y*ENHFH*[Z*l}*}
-gh]*JOGHKGJHJM*n}}8E8ā}}*[}8[y*EIMHFH*Y{y*l}*FFF
-gh]*JOGHKGJHJM*n}}8E8zāā~{y}*d}8Y*IMHFH*Y{y*l}*FFF
-gh]*JOGHKGJHJM*g}}8|8{}8]|}*]|}*MHHFH*Y{y*l}*FFF
-gh]*HIGHLGJHJM*n}}*[}8[y*QHHFH*}}*l}*āy}
-gh]*HIGHLGJHJM*g}}8|8{}8[}8}~y*[}8}~y*HFH*Y{y*l}*FFF
-gh]*HIGHLGJHJM*n}}8E8ā}}*]|}*EMHHFH*Y{y*l}*FFF
-gh]*HIGHLGJHJM*n}}8E8zāā~{y}*[}8}~y*MHHFH*Y{y*l}*FFF
-gh]*HIGHLGJHJM*g}}8ey|*ey|*LFH*Y{y*l}*FFF
-gh]*HJGHLGJHJM*}*d}8Y*EPQFH*[Z*l}*FFF8@y}A
-gh]*JHGHQGJHJM*Z}8c*[}8[y*EIIFH*y*l}*}
-Zm\*Y}y*JMHFH*]|}
-Zm\*}*JHHFH*[}8[y
-Zm\*āy}*KHHFH*d}8Y
-Zm\*FFF8@y}A*MHHHFH*[}8}~y
-Zm\*\}*IHHFH*[}8[y
-Zm\*k}*JMHFH*[}8[y
-Zm\*n}}}*JHHFH*[}8[y
