--- users/34567890.txt
@@ -1,13 +1,0 @@
-LY]*Lxvy}n)Y{rwlryju
-LY]*Ur{n})Mònuxyynvnw})M~{jkun
-LY]*Nyj{pwn
-XYN*9=89:8;9;=*Lx~{|n|*Lxvy}n)Y{rwlryju*@>7=*LK*]{~n*jurvnw}j}rxw
-XYN*9B89:8;9;=*Lxwln{}*Lxvy}n)Y{rwlryju*>>79*LK*]{~n*uxr|r{|
-XYN*:=89:8;9;=*Òyj{pwn*Lxvy}n)Y{rwlryju*;>979*_R[*]{~n*òyj{pwn
-XYN*:?89;8;9;>*|})junw}rw*Lxvy}n)Y{rwlryju*6>9979*r|j*]{~n*jurvnw}j}rxw
-XYN*9;89=8;9;>*X~n{}~{n)m~)lxvy}n)Nyj{pwn*Nyj{pwn*=>979*Jyyurlj}rxw*]{~n*777
-XYN*9;89=8;9;>*Mòsn~wn{*Lxvy}n)Y{rwlryju*6:@79*r|j*]{~n*jurvnw}j}rxw
-K^M*jurvnw}j}rxw*?9979*Lxvy}n)Y{rwlryju
-K^M*uxr|r{|*:>979*Lxvy}n)Y{rwlryju
-K^M*òyj{pwn*<>979*Ur{n})Mònuxyynvnw})M~{jkun
-K^M*n}nvnw}|*;>979*Lxvy}n)Y{rwlryju
