+++ users/41619558.txt
@@ -1,0 +1,10 @@
+KX\*Kwux|m(I
+WXM*:@7897:8::*kw}z{m{*Kwux|m(I*5@@6=*KPM*\z}m*{wz|qm{
+WXM*9A7897:8::*~qzmumv|*Kwux|m(I*:8>6@>*KPM*Nit{m*lq~mz{
+WXM*8:7897:8::*zm~mv}*Kwux|m(I*9<A68A*KPM*\z}m*lq~mz{
+WXM*9<7897:8::*kw}z{m{*Kwux|m(I*5@6<;*KJ*\z}m*itqumv|i|qwv
+J]L*I}|zm{*8*I}|zm{
+J]L*{iv|ñ*988*Kwux|m(I
+J]L*lq~mz{*=88*Kwux|m(I
+J]L*itqumv|i|qwv*=88*Kwux|m(I
+J]L*{wz|qm{*=88*Kwux|m(I
