+++ users/53739956.txt
@@ -1,0 +1,24 @@
+GTX*Gsqtxi$E
+GTX*Gsqtxi$F
+GTX*Gsqtxi$G
+STI*5934536466*wepemvi*Gsqtxi$G*57;2::*ZMV*Xvyi*xverwtsvx
+STI*4834536466*zmviqirx*Gsqtxi$E*;5286*ZMV*Xvyi*epmqirxexmsr
+STI*5534536466*gmriqe*Gsqtxi$G*1892:8*GLI*Xvyi*hmzivxmwwiqirx
+STI*6534536466*viqfsyvwiqirx*Gsqtxi$E*55<255*GF*Jepwi*hmzivw
+STI*6734536466*viqfsyvwiqirx*Gsqtxi$E*55:2;5*GF*Jepwi*xverwtsvx
+STI*5434536466*zmviqirx*Gsqtxi$G*5552;:*ZMV*Jepwi*hmzivxmwwiqirx
+STI*6934536466*wepemvi*Gsqtxi$E*54:2;*GLI*Xvyi*xverwtsvx
+STI*4634536466*gehiey*Gsqtxi$E*18;29;*GLI*Xvyi*hmzivxmwwiqirx
+STI*5734536466*wepemvi*Gsqtxi$E*9927=*GF*Xvyi*hmzivw
+STI*5734536466*zmviqirx*Gsqtxi$G*69;267*GLI*Jepwi*hmzivw
+STI*6734536466*tvmqi*Gsqtxi$F*65=25:*GLI*Jepwi*xverwtsvx
+STI*5734536466*zirxi*Gsqtxi$F*69928=*GF*Jepwi*wsvxmiw
+STI*5:34536466*tlevqegmi*Gsqtxi$E*1;82:<*ZMV*Jepwi*epmqirxexmsr
+STI*6834536466*gehiey*Gsqtxi$E*1<52=6*ZMV*Xvyi*epmqirxexmsr
+STI*6934536466*tlevqegmi*Gsqtxi$F*1929;*GLI*Jepwi*hmzivw
+STI*4534536466*viwxeyverx*Gsqtxi$E*1=<246*GF*Xvyi*epmqirxexmsr
+STI*5534536466*zmviqirx*Gsqtxi$E*6552<=*GF*Xvyi*hmzivw
+STI*4<34536466*tvmqi*Gsqtxi$F*6;92:6*GF*Xvyi*hmzivxmwwiqirx
+FYH*Eyxviw*4*Eyxviw
+FYH*hmzivxmwwiqirx*944*Gsqtxi$E
+FYH*wsvxmiw*544*Gsqtxi$G
