+++ users/55449072.txt
@@ -1,0 +1,19 @@
+Ubf*Uw2S
+abW*BFABCADBDD*uw*Uw2S*?KB@KH*UT*Xs~w*{w
+abW*CKABCADBDD*ww*Uw2S*CJF@CE*h[d*fw*v{w
+abW*ECABCADBDD*ww*Uw2S*CDD@IE*h[d*fw*s~{ws{
+abW*DIABCADBDD*{w*Uw2S*GH@CE*UT*Xs~w*s~{ws{
+abW*BGABCADBDD*ww*Uw2S*DEC@KC*UT*fw*v{w
+abW*CJABCADBDD*wss*Uw2S*?G@G*h[d*fw*{w
+abW*BEABCADBDD*ww*Uw2S*DFH@IE*UT*fw*{w
+abW*DCABCADBDD*zssu{w*Uw2S*?IJ@F*UZW*Xs~w*sû
+abW*BJABCADBDD*{w*Uw2S*DDC@GK*UT*Xs~w*s
+abW*CHABCADBDD*uw*Uw2S*?KI@BJ*UZW*Xs~w*s~{ws{
+abW*BJABCADBDD*ww*Uw2S*CKD@II*UT*Xs~w*{w
+abW*CGABCADBDD*u{ws*Uw2S*?EG@DE*UZW*fw*{w
+abW*BGABCADBDD*{ww*Uw2S*DKK@IF*UZW*Xs~w*{w
+abW*CFABCADBDD*wss*Uw2S*?FF@IH*h[d*fw*v{w
+abW*CJABCADBDD*zssu{w*Uw2S*?KG@JK*UT*Xs~w*sû
+abW*CJABCADBDD*wwuw*Uw2S*?GE@EK*UZW*Xs~w*sû
+TgV*Sw*B*Sw
+TgV*s*CBBB*Uw2S
