+++ users/67717699.txt
@@ -1,0 +1,23 @@
+O\`*O{y|q,M
+O\`*O{y|q,N
+[\Q*>C;<=;><>>*u~qyqz*O{y|q,N*>@?:<=*OTQ*Rmxq*puq~uqyqz
+[\Q*<?;<=;><>>*qzq*O{y|q,M*DE:A=*OTQ*`~q*puq~
+[\Q*><;<=;><>>*~qm~mz*O{y|q,M*9C?:=C*ON*`~q*~mz|{~
+[\Q*>B;<=;><>>*o{~q*O{y|q,N*9DB:A?*ON*`~q*puq~uqyqz
+[\Q*<=;<=;><>>*u~qyqz*O{y|q,N*==D:=D*OTQ*Rmxq*mxuyqzmu{z
+[\Q*=<;<=;><>>*u~qyqz*O{y|q,N*>>>:?@*OTQ*Rmxq*mzõ
+[\Q*<@;<=;><>>*ompqm*O{y|q,N*9CA:@@*OTQ*`~q*puq~
+[\Q*==;<=;><>>*|~uyq*O{y|q,N*=@<:>=*OTQ*Rmxq*~mz|{~
+[\Q*==;<=;><>>*mxmu~q*O{y|q,M*>>=:@B*ON*`~q*~mz|{~
+[\Q*<B;<=;><>>*~qqz*O{y|q,N*=?=:@C*OTQ*`~q*~mz|{~
+[\Q*=>;<=;><>>*u~qyqz*O{y|q,M*>=<:E?*ON*`~q*mxuyqzmu{z
+[\Q*=D;<=;><>>*~qqz*O{y|q,N*><C:<E*bU^*`~q*{~uq
+[\Q*<>;<=;><>>*o{~q*O{y|q,M*9BC:=@*ON*Rmxq*mxuyqzmu{z
+[\Q*<A;<=;><>>*qzq*O{y|q,M*=@E:AD*ON*Rmxq*~mz|{~
+[\Q*<@;<=;><>>*u~qyqz*O{y|q,M*>EA:CE*ON*Rmxq*puq~uqyqz
+[\Q*?=;<=;><>>*|~uyq*O{y|q,N*>?D:EC*ON*`~q*~mz|{~
+NaP*M~q*<*M~q
+NaP*mzõ*><<*O{y|q,M
+NaP*{~uq*=<<<*O{y|q,M
+NaP*puq~*><<*O{y|q,N
+NaP*~mz|{~*A<<*O{y|q,N
