+++ users/71030817.txt
@@ -1,0 +1,26 @@
+GTX*Gsqtxi$E
+GTX*Gsqtxi$F
+GTX*Gsqtxi$G
+STI*4734536466*tvmqi*Gsqtxi$E*6=:2;=*GF*Xvyi*werxí
+STI*4:34536466*viziry*Gsqtxi$E*6672;;*GF*Xvyi*epmqirxexmsr
+STI*4:34536466*zirxi*Gsqtxi$F*6;52=8*GF*Xvyi*hmzivw
+STI*4;34536466*zmviqirx*Gsqtxi$F*55;26*ZMV*Xvyi*werxí
+STI*4<34536466*viqfsyvwiqirx*Gsqtxi$F*65:249*GLI*Xvyi*xverwtsvx
+STI*4=34536466*gehiey*Gsqtxi$F*1552;=*ZMV*Xvyi*werxí
+STI*5434536466*zirxi*Gsqtxi$E*68=2::*ZMV*Xvyi*hmzivw
+STI*5434536466*gmriqe*Gsqtxi$G*1<4257*GLI*Xvyi*hmzivw
+STI*5634536466*tlevqegmi*Gsqtxi$E*17;285*GF*Xvyi*epmqirxexmsr
+STI*5834536466*gehiey*Gsqtxi$F*158249*GLI*Xvyi*xverwtsvx
+STI*5;34536466*zirxi*Gsqtxi$G*5:8246*GF*Xvyi*hmzivw
+STI*5;34536466*wepemvi*Gsqtxi$F*6=927<*GF*Xvyi*wsvxmiw
+STI*5=34536466*wepemvi*Gsqtxi$F*6:;2=7*ZMV*Xvyi*werxí
+STI*6434536466*gmriqe*Gsqtxi$G*15;247*GLI*Xvyi*hmzivxmwwiqirx
+STI*6834536466*iwwirgi*Gsqtxi$E*18924:*GF*Xvyi*werxí
+STI*6;34536466*zirxi*Gsqtxi$G*5:9246*ZMV*Xvyi*xverwtsvx
+STI*6;34536466*tlevqegmi*Gsqtxi$G*166296*GF*Xvyi*wsvxmiw
+STI*6;34536466*zirxi*Gsqtxi$G*66727*GLI*Xvyi*hmzivxmwwiqirx
+STI*7534536466*tvmqi*Gsqtxi$E*6=5297*GLI*Xvyi*hmzivxmwwiqirx
+FYH*Eyxviw*424*Eyxviw
+FYH*wsvxmiw*64424*Gsqtxi$G
+FYH*epmqirxexmsr*544424*Gsqtxi$E
+FYH*hmzivxmwwiqirx*544424*Gsqtxi$E
