+++ users/72432105.txt
@@ -1,0 +1,10 @@
+HUY*Htruyj%F
+HUY*Htruyj%G
+HUY*Htruyj%H
+TUJ*6;45647577*umfwrfhnj*Htruyj%G*2<:37*HG*Ywzj*in{jwx
+TUJ*8545647577*wjxyfzwfsy*Htruyj%F*2>35:*[NW*Kfqxj*xtwynjx
+TUJ*5>45647577*{jsyj*Htruyj%F*6=:3<9*HG*Ywzj*ywfsxutwy
+TUJ*6:45647577*wjrgtzwxjrjsy*Htruyj%F*>53:*[NW*Ywzj*in{jwx
+GZI*Fzywjx*5*Fzywjx
+GZI*in{jwynxxjrjsy*:55*Htruyj%F
+GZI*ywfsxutwy*6555*Htruyj%H
