+++ users/79748315.txt
@@ -1,0 +1,21 @@
+Yfj*Y{6W
+Yfj*Y{6X
+Yfj*Y{6Y
+Yfj*Y{6[w}{
+ef[*FHEFGEHFHH*{{*Y{6W*GFKDNJ*YX*j{*{
+ef[*FHEFGEHFHH*{x{{*Y{6W*GMIDIG*l_h*j{*z{{{
+ef[*FNEFGEHFHH*~wwy{*Y{6X*CKNDL*Y^[*j{*z{{{
+ef[*FNEFGEHFHH*{{*Y{6W*HFIDGM*l_h*j{*{
+ef[*FOEFGEHFHH*{{*Y{6Y*HFLDLK*l_h*j{*z{{{
+ef[*FOEFGEHFHH*z{{*Y{6W*CHFDF*Y^[*j{*wÿ
+ef[*GJEFGEHFHH*{*Y{6W*HGIDJM*Y^[*j{*wÿ
+ef[*GLEFGEHFHH*{{*Y{6W*GNFDHG*l_h*j{*w
+ef[*HOEFGEHFHH*y{*Y{6Y*COHDKG*Y^[*j{*w
+ef[*HFEFIEHFHH*{{y{*Y{6W*CMFDKJ*YX*j{*w
+ef[*FJEFJEHFHK*l{{6C6ÿ{{*Y{6W*CLFDF*Wyw*j{*DDD
+ef[*FJEFJEHFHK*l{{6C6xÿÿ|yw{*Y{6X*LFDF*Wyw*j{*DDD
+ef[*FJEFJEHFHK*e{{6z6y{6Y{6[w}{*Y{6[w}{*IFFDF*Wyw*j{*DDD
+XkZ*W{*FDF*W{
+XkZ*wÿ*KFFDF*Y{6W
+XkZ*w{w*KFFDF*Y{6X
+XkZ*w*IFFDF*Y{6W
