+++ users/84162237.txt
@@ -1,0 +1,13 @@
+O\`*O{y|q,M
+O\`*O{y|q,N
+O\`*O{y|q,O
+[\Q*=@;<=;><>>*~qyn{~qyqz*O{y|q,N*C<:BE*bU^*Rmxq*mzõ
+[\Q*>?;<=;><>>*u~qyqz*O{y|q,O*EA:@@*OTQ*Rmxq*mxuyqzmu{z
+[\Q*>E;<=;><>>*o{~q*O{y|q,N*9@C:>*bU^*`~q*puq~uqyqz
+[\Q*<?;<=;><>>*qqzoq*O{y|q,O*9C>:CE*OTQ*`~q*mxuyqzmu{z
+[\Q*>D;<=;><>>*ouzqym*O{y|q,M*9AC:A*ON*Rmxq*{~uq
+NaP*M~q*<*M~q
+NaP*~mz|{~*A<<*O{y|q,N
+NaP*{~uq*=<<*O{y|q,O
+NaP*mxuyqzmu{z*?<<*O{y|q,N
+NaP*puq~uqyqz*?<<*O{y|q,N
