--- users/87654321.txt
@@ -1,8 +1,0 @@
-R_c*R~|t/R~p}
-R_c*R~|t/Øpv}t
-^_T*?B>?@>A?AC*Tt}rt*R~|t/R~p}*E?=?*RQ*ct*p}~
-^_T*?F>?@>A?AC*P|p~}/Prwp*R~|t/R~p}*GH=HH*RQ*ct*w~x}v
-^_T*@A>?@>A?AC*ext|t}/Øpv}t*R~|t/R~p}*@D?=?*eXa*ct*øpv}t
-QdS*p}~*A??=?*R~|t/R~p}
-QdS*w~x}v*B??=?*R~|t/R~p}
-QdS*øpv}t*C??=?*R~|t/Øpv}t
