+++ users/92159840.txt
@@ -1,0 +1,16 @@
+LY]*Lxvy}n)J
+LY]*Lxvy}n)K
+XYN*;>89:8;9;;*r{nvnw}*Lxvy}n)K*;9:79?*LQN*]{~n*jurvnw}j}rxw
+XYN*:=89:8;9;;*{nvkx~{|nvnw}*Lxvy}n)J*A:7=@*LQN*Oju|n*jurvnw}j}rxw
+XYN*:@89:8;9;;*nw}n*Lxvy}n)J*@=7?;*LQN*Oju|n*jurvnw}j}rxw
+XYN*9<89:8;9;;*yqj{vjlrn*Lxvy}n)K*6@97;:*LQN*Oju|n*|jw}ò
+XYN*:@89:8;9;;*|jujr{n*Lxvy}n)J*;:A7=<*LK*Oju|n*}{jw|yx{}
+XYN*9@89:8;9;;*y{rvn*Lxvy}n)K*;AB7??*LK*]{~n*}{jw|yx{}
+XYN*9?89:8;9;;*r{nvnw}*Lxvy}n)K*:AA7:<*_R[*Oju|n*|x{}rn|
+XYN*:=89:8;9;;*yqj{vjlrn*Lxvy}n)J*6><7;*_R[*]{~n*|jw}ò
+XYN*::89:8;9;;*r{nvnw}*Lxvy}n)J*;;:7>B*_R[*Oju|n*|jw}ò
+XYN*9<89:8;9;;*ljmnj~*Lxvy}n)K*6B=7<;*LK*Oju|n*|jw}ò
+K^M*J~}{n|*9*J~}{n|
+K^M*mrn{|*:999*Lxvy}n)J
+K^M*jurvnw}j}rxw*;99*Lxvy}n)J
+K^M*}{jw|yx{}*>99*Lxvy}n)J
