+++ users/96578654.txt
@@ -1,0 +1,20 @@
+IVZ*Iusvzk&G
+UVK*8<56758688*igjkg{*Iusvzk&G*3<=4;=*\OX*Zx{k*jo|kxy
+UVK*6=56758688*xkyzg{xgtz*Iusvzk&G*39649=*\OX*Zx{k*jo|kxy
+UVK*6?56758688*vxosk*Iusvzk&G*87746<*\OX*Zx{k*ygtzï
+UVK*7<56758688*vxosk*Iusvzk&G*76<4;*IH*Lgryk*ygtzï
+UVK*9756758688*iu{xyky*Iusvzk&G*3?<46;*INK*Zx{k*yuxzoky
+UVK*7756758688*vngxsgiok*Iusvzk&G*37848=*\OX*Zx{k*zxgtyvuxz
+UVK*8;56758688*xk|kt{*Iusvzk&G*8=94>9*IH*Lgryk*zxgtyvuxz
+UVK*8<56758688*xk|kt{*Iusvzk&G*8<7497*INK*Lgryk*grosktzgzout
+UVK*8756758688*xk|kt{*Iusvzk&G*><498*INK*Lgryk*zxgtyvuxz
+UVK*6756758688*|ktzk*Iusvzk&G*?;479*INK*Lgryk*grosktzgzout
+UVK*6<56758688*iu{xyky*Iusvzk&G*3;748?*INK*Zx{k*jo|kxy
+UVK*8956758688*igjkg{*Iusvzk&G*38:478*IH*Lgryk*jo|kxy
+UVK*7=56758688*ygrgoxk*Iusvzk&G*88=46?*IH*Lgryk*grosktzgzout
+UVK*7?56758688*vngxsgiok*Iusvzk&G*3894<9*IH*Lgryk*zxgtyvuxz
+UVK*8=56758688*xk|kt{*Iusvzk&G*88:4<>*\OX*Zx{k*jo|kxzoyyksktz
+UVK*7656758688*xk|kt{*Iusvzk&G*>74;?*INK*Zx{k*grosktzgzout
+UVK*6856758688*vngxsgiok*Iusvzk&G*3:=47>*\OX*Lgryk*jo|kxy
+H[J*G{zxky*6*G{zxky
+H[J*yuxzoky*7666*Iusvzk&G
